change repo folder structure, add first part of documentation and removing unused code
