+++ code/seqtab_stats.R
